change watch to listen
Listen will insert the restart at the very end of a state run so that
everything runs sequentially.  If you have a watch that is watching
multiple things, the service gets restarted any time any of the configs
get changed.  If you use listen, the restart gets tacked onto the end of
the list, so it only runs one restart at the very end of the highstate.

closes #101